Moved mysql test setup.
